Fix constraints inside stack view to prevent conflict
When an element in a UIStackView is hidden, a constraint of H:[element]0 is set (element width set to 0).
Since there was an explicit width constraint already, these two were in conflict.  By reducing the explicit width
constraint's priority the conflict is avoided, allowing UISV to hide the element how it likes.