Update equality for WistiaMedia, WistiaMediaEmbedOptions
Object equality (==) now goes deeper than just the hashedID.  If meaningful aspect of a Media are different,
it is no longer equivalent to it's old self.  In this case "meaningful" refers to qualities that affect playback,
look, and feel.   The soft stuff of humans vs the hard numbers of machines.